[#39] feat: Hitter 컴포넌트 분리
- 재사용성을 위해 Hitter 컴포넌트를 따로 분리함
