+++ FE/src/components/GamePage/BaseballField/Field/Hitter/hitter.png
